start-data-viz - break each specific chart out into its own component while still doing data acquisition at topp level App component
