+++ app/package.json
@@ -9,7 +9,8 @@
     "react": "^16.7.0",
     "react-dom": "^16.7.0",
     "react-router-dom": "^4.3.1",
-    "react-scripts": "2.1.3"
+    "react-scripts": "2.1.3",
+    "react-splitter-layout": "^3.0.1"
   },
   "scripts": {
     "start": "react-scripts start",
